Split ProductRowImporter into tagged per-concern row steps
ProductRowStepInterface (tag core.import.product_row_step, autoconfigured
via PrestaShopExtension, priority-sorted, module-extensible with a low BC
promise) + AbstractProductRowStep and ProductRowValueTrait hosting the
shared helpers. The 13 appliers become one step service each (verbatim
moves, &$messages converted to returned lists, coalesced per batch by
PhaseBatchResult); ProductFieldsStep absorbs the manufacturer resolution
and the tax-divisor memo. The orchestrator keeps the identity flow:
language check, findRowMatch and its two defenses, update-vs-create,
the supports/apply loop and the row catch-all (1261 -> ~180 lines,
22 -> 9 constructor dependencies). Step order lives in one commented
block of import_engine.yml (spaced priorities 1300->100).
