+++ src/PrestaShopBundle/DependencyInjection/PrestaShopExtension.php
@@ -9,6 +9,8 @@ namespace PrestaShopBundle\DependencyInjection;
 use PrestaShop\PrestaShop\Adapter\Configuration;
 use PrestaShop\PrestaShop\Core\ConfigurationInterface;
 use PrestaShop\PrestaShop\Core\Http\CookieOptions;
+use PrestaShop\PrestaShop\Core\Import\Engine\EntityImporter\Product\ProductRowImporter;
+use PrestaShop\PrestaShop\Core\Import\Engine\EntityImporter\Product\Step\ProductRowStepInterface;
 use PrestaShop\PrestaShop\Core\Import\Engine\EntityImporterInterface;
 use PrestaShop\PrestaShop\Core\Import\Engine\EntityImporterRegistry;
 use PrestaShop\PrestaShop\Core\Security\OAuth2\AuthorisationServerInterface;
@@ -45,6 +47,13 @@ class PrestaShopExtension extends Extension implements PrependExtensionInterface
         // long as their definitions enable autoconfiguration)
         $container->registerForAutoconfiguration(EntityImporterInterface::class)
             ->addTag(EntityImporterRegistry::SERVICE_TAG)
+        ;
+
+        // Automatically tag product row steps (same mechanism: module services
+        // are collected too; an autoconfigured step without an explicit tag
+        // priority runs at 0, after every core step)
+        $container->registerForAutoconfiguration(ProductRowStepInterface::class)
+            ->addTag(ProductRowImporter::STEP_TAG)
         ;
     }
